requirements.txt
Django==1.11.2
django-extensions==1.8.0
beautifulsoup4==4.6.0
django-prettyjson==0.3.0
